+++ pytorch_forecasting/models/tcn/__init__.py
@@ -1,22 +1,27 @@
-''' TCN With Covariates'''
+""" TCN With Covariates"""
+
+from typing import Dict, List, Tuple
 
 import torch
 import torch.nn as nn
 from torch.nn.utils import weight_norm
-from typing import Dict, List, Tuple
 
-from pytorch_forecasting.models.nn import MultiEmbedding
+from pytorch_forecasting.data.timeseries import TimeSeriesDataSet
 from pytorch_forecasting.models import BaseModelWithCovariates
+from pytorch_forecasting.models.nn import MultiEmbedding
 
 
 class GAP1d(nn.Module):
     "Global Adaptive Pooling + Flatten"
+
     def __init__(self, output_size=1):
-        super(GAP1d,self).__init__()
+        super(GAP1d, self).__init__()
         self.gap = nn.AdaptiveAvgPool1d(output_size)
+
     def forward(self, x):
-        res=self.gap(x)
-        return res.view(res.size(0), -1) ##fastai's layer has problems with latest torch
+        res = self.gap(x)
+        return res.view(res.size(0), -1)  # fastai's layer has problems with latest torch
+
 
 class Chomp1d(nn.Module):
     def __init__(self, chomp_size):
@@ -24,32 +29,39 @@ class Chomp1d(nn.Module):
         self.chomp_size = chomp_size
 
     def forward(self, x):
-        return x[:, :, :-self.chomp_size].contiguous()
+        return x[:, :, : -self.chomp_size].contiguous()
+
 
 class Flatten(nn.Module):
-  def __init__(self,):
-    super(Flatten,self).__init__()
-  def forward(self,x):
-    return x.view(x.size(0),-1)
+    def __init__(
+        self,
+    ):
+        super(Flatten, self).__init__()
+
+    def forward(self, x):
+        return x.view(x.size(0), -1)
 
 
 class TemporalBlock(nn.Module):
     def __init__(self, n_inputs, n_outputs, kernel_size, stride, dilation, padding, dropout=0.2):
         super(TemporalBlock, self).__init__()
-        self.conv1 = weight_norm(nn.Conv1d(n_inputs, n_outputs, kernel_size,
-                                           stride=stride, padding=padding, dilation=dilation))
+        self.conv1 = weight_norm(
+            nn.Conv1d(n_inputs, n_outputs, kernel_size, stride=stride, padding=padding, dilation=dilation)
+        )
         self.chomp1 = Chomp1d(padding)
         self.relu1 = nn.ReLU()
         self.dropout1 = nn.Dropout(dropout)
 
-        self.conv2 = weight_norm(nn.Conv1d(n_outputs, n_outputs, kernel_size,
-                                           stride=stride, padding=padding, dilation=dilation))
+        self.conv2 = weight_norm(
+            nn.Conv1d(n_outputs, n_outputs, kernel_size, stride=stride, padding=padding, dilation=dilation)
+        )
         self.chomp2 = Chomp1d(padding)
         self.relu2 = nn.ReLU()
         self.dropout2 = nn.Dropout(dropout)
 
-        self.net = nn.Sequential(self.conv1, self.chomp1, self.relu1, self.dropout1,
-                                 self.conv2, self.chomp2, self.relu2, self.dropout2)
+        self.net = nn.Sequential(
+            self.conv1, self.chomp1, self.relu1, self.dropout1, self.conv2, self.chomp2, self.relu2, self.dropout2
+        )
         self.downsample = nn.Conv1d(n_inputs, n_outputs, 1) if n_inputs != n_outputs else None
         self.relu = nn.ReLU()
         self.init_weights()
@@ -62,7 +74,7 @@ class TemporalBlock(nn.Module):
 
     def forward(self, x):
         out = self.net(x)
-        res = x if self.downsample is None else self.downsample(x) ##保持时序的长度不变，但是内部的特征会发生改变
+        res = x if self.downsample is None else self.downsample(x)  # 保持时序的长度不变，但是内部的特征会发生改变
         return self.relu(out + res)
 
 
@@ -73,10 +85,19 @@ class TemporalConvNet(nn.Module):
         num_levels = len(num_channels)
         for i in range(num_levels):
             dilation_size = 2 ** i
-            in_channels = num_inputs if i == 0 else num_channels[i-1]
+            in_channels = num_inputs if i == 0 else num_channels[i - 1]
             out_channels = num_channels[i]
-            layers += [TemporalBlock(in_channels, out_channels, kernel_size, stride=1, dilation=dilation_size,
-                                     padding=(kernel_size-1) * dilation_size, dropout=dropout)]
+            layers += [
+                TemporalBlock(
+                    in_channels,
+                    out_channels,
+                    kernel_size,
+                    stride=1,
+                    dilation=dilation_size,
+                    padding=(kernel_size - 1) * dilation_size,
+                    dropout=dropout,
+                )
+            ]
 
         self.network = nn.Sequential(*layers)
 
@@ -84,18 +105,15 @@ class TemporalConvNet(nn.Module):
         return self.network(x)
 
 
-
-
 class TCN(BaseModelWithCovariates):
-
     def __init__(
         self,
-        input_size: int, 
+        input_size: int,
         output_size: int,
         n_hidden_layers: List[int],
-        conv_dropout:float,
-        fc_dropout:float,
-        kernel_size:int,
+        conv_dropout: float,
+        fc_dropout: float,
+        kernel_size: int,
         x_reals: List[str],
         x_categoricals: List[str],
         embedding_sizes: Dict[str, Tuple[int, int]],
@@ -108,7 +126,6 @@ class TCN(BaseModelWithCovariates):
         time_varying_reals_decoder: List[str],
         embedding_paddings: List[str],
         categorical_groups: Dict[str, List[str]],
-        
         **kwargs,
     ):
         """
@@ -116,20 +133,20 @@ class TCN(BaseModelWithCovariates):
 
         Based on the article
         `TCN: An Empirical Evaluation of Generic Convolutional and Recurrent Networks for Sequence Modeling
-       <https://arxiv.org/abs/1803.01271>`_. TCN has a better performance than LSTM in time series forecasting with much faster training speed.
+        <https://arxiv.org/abs/1803.01271>`_. TCN has a better performance than LSTM in time series forecasting
+        with much faster training speed.
 
         Args:
-            n_hidden_layers:the number of temporal block.In this implementation,every temporal block has fixed three casual dialation cnn kernel according to the original paper
-            conv_dropout: dropout rate for the output of every temporal block
-            fc_dropout: In this implementation,we use pooling and flatten(Gap1D) to make the output into 1-d dimension and use a linear layer to complete time series forecasting tasks 
-            fc_drouput is the droup out rate for the output of Gap1D
-            kernel_size: the kernel size of the cnn
-            prediction_length: Length of the prediction. Also known as 'horizon'.
-            context_length: Number of time units that condition the predictions. Also known as 'lookback period'.
+            n_hidden_layers (int): the number of temporal blocks. In this implementation,every temporal block has fixe
+                three casual dialation cnn kernel according to the original paper
+            conv_dropout (float): dropout rate for the output of every temporal block
+            fc_dropout (float): In this implementation, we use pooling and flatten(Gap1D) to make the output into 1D
+                dimension and use a linear layer to complete time series forecasting tasks
+                fc_drouput is the droup out rate for the output of Gap1D
+            kernel_size (int): the kernel size of the cnn
+            prediction_length (int): Length of the prediction. Also known as 'horizon'.
+            context_length (int): Number of time units that condition the predictions. Also known as 'lookback period'.
                 Should be between 1-10 times the prediction length.
-            backcast_loss_ratio: weight of backcast in comparison to forecast when calculating the loss.
-                A weight of 1.0 means that forecast and backcast loss is weighted the same (regardless of backcast and
-                forecast lengths). Defaults to 0.0, i.e. no weight.
             loss: loss to optimize. Defaults to MASE().
             log_gradient_flow: if to log gradient flow, this takes time and should be only done to diagnose training
                 failures
@@ -141,7 +158,7 @@ class TCN(BaseModelWithCovariates):
 
         # saves arguments in signature to `.hparams` attribute, mandatory call - do not skip this
         self.save_hyperparameters()
-        
+
         # pass additional arguments to BaseModel.__init__, mandatory call - do not skip this
         super().__init__(**kwargs)
 
@@ -160,22 +177,20 @@ class TCN(BaseModelWithCovariates):
         ) + len(self.reals)
 
         # create network that will be fed with continious variables and embeddings
-        self.network = nn.Sequential(TemporalConvNet(
-            num_inputs=n_features,
-            num_channels=self.hparams.n_hidden_layers,
-            kernel_size=self.hparams.kernel_size,
-            dropout=self.hparams.conv_dropout),
+        self.network = nn.Sequential(
+            TemporalConvNet(
+                num_inputs=n_features,
+                num_channels=self.hparams.n_hidden_layers,
+                kernel_size=self.hparams.kernel_size,
+                dropout=self.hparams.conv_dropout,
+            ),
             GAP1d(),
             nn.Dropout(self.hparams.fc_dropout),
-            nn.Linear(self.hparams.n_hidden_layers[-1],self.hparams.output_size)
-            
-            
+            nn.Linear(self.hparams.n_hidden_layers[-1], self.hparams.output_size),
         )
-
 
     def forward(self, x: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
         # x is a batch generated based on the TimeSeriesDataset
-        batch_size = x["encoder_lengths"].size(0)
         embeddings = self.input_embeddings(x["encoder_cat"])  # returns dictionary with embedding tensors
         network_input = torch.cat(
             [x["encoder_cont"]]
@@ -186,15 +201,13 @@ class TCN(BaseModelWithCovariates):
             ],
             dim=-1,
         )
-        prediction = self.network(network_input.permute(0,2,1))
-        
+        prediction = self.network(network_input.permute(0, 2, 1))
 
         # We need to return a dictionary that at least contains the prediction and the target_scale.
         # The parameter can be directly forwarded from the input.
-        #return dict(prediction=prediction, target_scale=x["target_scale"])
+        # return dict(prediction=prediction, target_scale=x["target_scale"])
         prediction = self.transform_output(prediction, target_scale=x["target_scale"])
         return self.to_network_output(prediction=prediction)
-
 
     @classmethod
     def from_dataset(cls, dataset: TimeSeriesDataSet, **kwargs):
@@ -208,4 +221,3 @@ class TCN(BaseModelWithCovariates):
         assert dataset.min_encoder_length == dataset.max_encoder_length, "Encoder only supports a fixed length"
 
         return super().from_dataset(dataset, **new_kwargs)
-